Autonomtlägeknapp och tidtagning testat och klart.
