+++ Telemetry/Telemetry.cs
@@ -14,20 +14,19 @@ namespace Telemetry
         public static byte LEFT = 0x03;
         public static byte RIGHT = 0x04;
         public static byte STOP = 0x05;
-        public static byte START_STOP_TIMER = 0x06;
+        public static byte STOP_TIMER = 0x06;
         public static byte WAIT_FOR_P = 0x07;
         public static byte WAIT_FOR_D = 0x08;
+        public static byte START_TIMER = 0x09;
 
         private SerialPort connection;
         private long startTime;
-        private bool counting;
 
         private System.Text.StringBuilder log = new System.Text.StringBuilder();
         private int counter;
 
         public Serial(String port, Int32 baud, Parity p, Int32 databits, StopBits b)
         {
-            counting = false;
             connection = new SerialPort(port, baud, p, databits, b);
             connection.DataReceived += new SerialDataReceivedEventHandler(_serialPort_DataReceived); 
             if (!connection.IsOpen)
@@ -76,21 +75,17 @@ namespace Telemetry
             //Append to log stringbuilder 
             for (int i = 0; i < bytesRead; i++)
             {
-                if(buffer[i] == Serial.START_STOP_TIMER)
+                if (buffer[i] == Serial.STOP_TIMER)
                 {
-                    if (counting)
-                    {
-                        counting = false;
-                        Console.WriteLine("");
-                        Console.Write("Finished in: ");
-                        Console.Write(DateTime.Now.Ticks / TimeSpan.TicksPerMillisecond - startTime);
-                        Console.WriteLine(" milliseconds");
-                    }
-                    else
-                    {
-                        counting = true;
-                        startTime = DateTime.Now.Ticks / TimeSpan.TicksPerMillisecond;
-                    }
+                    Console.WriteLine("");
+                    Console.Write("Finished in: ");
+                    Console.Write(DateTime.Now.Ticks / TimeSpan.TicksPerMillisecond - startTime);
+                    Console.WriteLine(" milliseconds");
+                    Console.Write("Press a key>> ");
+                }
+                else if (buffer[i] == Serial.START_TIMER)
+                {
+                    startTime = DateTime.Now.Ticks / TimeSpan.TicksPerMillisecond;
                 }
                 log.Append(counter.ToString()+":\t0b"+Convert.ToString(buffer[i], 2).PadLeft(8, '0')+"\t0x"+Convert.ToString(buffer[i], 16).PadLeft(2, '0')+"\n");
                 counter++;
